+++ kurbo/src/axis.rs
@@ -15,7 +15,7 @@ pub enum Axis {
 impl Axis {
     /// Get the axis perpendicular to this one.
     #[inline]
-    pub fn cross(self) -> Self {
+    pub const fn cross(self) -> Self {
         match self {
             Self::Horizontal => Self::Vertical,
             Self::Vertical => Self::Horizontal,
@@ -27,8 +27,11 @@ impl Axis {
     /// The axis value is the one matching the axis (e.g. `y` for [`Self::Vertical`]).
     /// The cross value is the other one.
     #[inline]
-    pub fn pack_point(self, axis_value: f64, cross_value: f64) -> Point {
-        self.pack_xy(axis_value, cross_value).into()
+    pub const fn pack_point(self, axis_value: f64, cross_value: f64) -> Point {
+        match self {
+            Self::Horizontal => Point::new(axis_value, cross_value),
+            Self::Vertical => Point::new(cross_value, axis_value),
+        }
     }
 
     /// Create a new [`Size`] by arranging the given magnitudes.
@@ -36,8 +39,11 @@ impl Axis {
     /// The axis value is the one matching the axis (e.g. `height` for [`Self::Vertical`]).
     /// The cross value is the other one.
     #[inline]
-    pub fn pack_size(self, axis_value: f64, cross_value: f64) -> Size {
-        self.pack_xy(axis_value, cross_value).into()
+    pub const fn pack_size(self, axis_value: f64, cross_value: f64) -> Size {
+        match self {
+            Self::Horizontal => Size::new(axis_value, cross_value),
+            Self::Vertical => Size::new(cross_value, axis_value),
+        }
     }
 
     /// Create a new [`Vec2`] by arranging the given magnitudes.
@@ -45,8 +51,11 @@ impl Axis {
     /// The axis value is the one matching the axis (e.g. `y` for [`Self::Vertical`]).
     /// The cross value is the other one.
     #[inline]
-    pub fn pack_vec2(self, axis_value: f64, cross_value: f64) -> Vec2 {
-        self.pack_xy(axis_value, cross_value).into()
+    pub const fn pack_vec2(self, axis_value: f64, cross_value: f64) -> Vec2 {
+        match self {
+            Self::Horizontal => Vec2::new(axis_value, cross_value),
+            Self::Vertical => Vec2::new(cross_value, axis_value),
+        }
     }
 
     /// Create a new `(x, y)` pair by arranging the given magnitudes.
@@ -54,7 +63,7 @@ impl Axis {
     /// The axis value is the one matching the axis (e.g. `y` for [`Self::Vertical`]).
     /// The cross value is the other one.
     #[inline]
-    pub fn pack_xy(self, axis_value: f64, cross_value: f64) -> (f64, f64) {
+    pub const fn pack_xy(self, axis_value: f64, cross_value: f64) -> (f64, f64) {
         match self {
             Self::Horizontal => (axis_value, cross_value),
             Self::Vertical => (cross_value, axis_value),