+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This release has an [MSRV][] of 1.85.
 ## Changed
 
 - Improve performance of `RoundedRect::winding` and `RoundedRect::contains`. ([#534][] by [@tomcur][])
+- `Axis`, `Insets`, `Line::reversed`, `Rect::min_x/max_x/min_y/max_y`, `RoundedRectRadii::abs/clamp/as_single_radius`, `RoundedRect::width/height/radii/rect/origin/center`, `Size::aspect_ratio`, `Stroke::with_join/with_miter_limit/with_start_cap/with_end_cap/with_caps` are all now `const`. ([#536][] by [@xStrom])
 
 ## [0.13.0] (2025-11-27)
 
@@ -193,6 +194,7 @@ Note: A changelog was not kept for or before this release
 [@tomcur]: https://github.com/tomcur
 [@waywardmonkeys]: https://github.com/waywardmonkeys
 [@xorgy]: https://github.com/xorgy
+[@xStrom]: https://github.com/xStrom
 
 [#288]: https://github.com/linebender/kurbo/pull/288
 [#334]: https://github.com/linebender/kurbo/pull/334
@@ -268,6 +270,7 @@ Note: A changelog was not kept for or before this release
 [#526]: https://github.com/linebender/kurbo/pull/526
 [#527]: https://github.com/linebender/kurbo/pull/527
 [#534]: https://github.com/linebender/kurbo/pull/534
+[#536]: https://github.com/linebender/kurbo/pull/536
 
 [Unreleased]: https://github.com/linebender/kurbo/compare/v0.13.0...HEAD
 [0.13.0]: https://github.com/linebender/kurbo/releases/tag/v0.13.0